Split out interfaces into modules
