+++ src/www/api.ts
@@ -2,9 +2,9 @@
 // Core API
 export {
     Inset,
-    IInsetAPI,
-    IInset,
-    IInsetCallbackFunc,
-    IInsetConfiguration
+    IInsetAPI
 } from './Inset';
+export {IInset} from './IInset';
 export {InsetMask} from './InsetMask';
+export {IInsetCallbackFunc} from './IInsetCallbackFunc';
+export {IInsetConfiguration} from './IInsetConfiguration';
